--- packages/esm-appointments-app/__mocks__/patient-appointments.mock.ts
@@ -1,94 +1,0 @@
-export const mockAppointments = {
-  pastAppointments: [
-    {
-      uuid: 'a4662406-2bfe-49be-8465-6dbdfe9b4d5d',
-      appointmentNumber: '0000',
-      patient: {
-        identifier: '1003R8',
-        name: 'Brian Evans',
-        uuid: '8f75dbf1-833d-498a-bc99-3a24cad5cee9',
-      },
-      service: {
-        appointmentServiceId: 1,
-        name: 'HIV return visit',
-        description: null,
-        speciality: {},
-        startTime: '',
-        endTime: '',
-        maxAppointmentsLimit: null,
-        durationMins: null,
-        location: {},
-        uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-        color: '#006400',
-        initialAppointmentStatus: 'Scheduled',
-        creatorName: null,
-      },
-      serviceType: null,
-      provider: null,
-      location: {
-        name: 'Outpatient Clinic',
-        uuid: 'f76c0c8e-2c3a-443c-b26d-96a9f3847764',
-      },
-      startDateTime: '2021-10-28T11:14:17.064Z',
-      endDateTime: 1649935800000,
-      appointmentKind: 'Scheduled',
-      status: 'Scheduled',
-      comments: null,
-      additionalInfo: null,
-      teleconsultation: null,
-      providers: [],
-      voided: false,
-      extensions: {
-        patientEmailDefined: false,
-      },
-      teleconsultationLink: null,
-      recurring: false.valueOf,
-    },
-  ],
-  upcomingAppointments: [
-    {
-      uuid: 'a4662406-2bfe-49be-8465-6dbdfe9b4d5d',
-      appointmentNumber: '0000',
-      patient: {
-        identifier: '1003R8',
-        name: 'Brian Evans',
-        uuid: '8f75dbf1-833d-498a-bc99-3a24cad5cee9',
-      },
-      service: {
-        appointmentServiceId: 1,
-        name: 'Clinical observation',
-        description: null,
-        speciality: {},
-        startTime: '',
-        endTime: '',
-        maxAppointmentsLimit: null,
-        durationMins: null,
-        location: {},
-        uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-        color: '#006400',
-        initialAppointmentStatus: 'Scheduled',
-        creatorName: null,
-      },
-      serviceType: null,
-      provider: null,
-      location: {
-        name: 'Outpatient Clinic',
-        uuid: 'f76c0c8e-2c3a-443c-b26d-96a9f3847764',
-      },
-      startDateTime: 1649934000000,
-      endDateTime: 1649935800000,
-      appointmentKind: 'Scheduled',
-      status: 'Scheduled',
-      comments: null,
-      additionalInfo: null,
-      teleconsultation: null,
-      providers: [],
-      voided: false,
-      extensions: {
-        patientEmailDefined: false,
-      },
-      teleconsultationLink: null,
-      recurring: false.valueOf,
-    },
-  ],
-};
