+++ packages/esm-appointments-app/src/patient-search/patient-search.component.tsx
@@ -1,10 +1,9 @@
 import React, { useState } from 'react';
-import { useTranslation } from 'react-i18next';
+
 import BasicSearch from './basic-search.component';
 import AdvancedSearch from './advanced-search.component';
 import PatientScheduledVisits from './patient-scheduled-visits.component';
 import SearchResults from './search-results.component';
-import PatientForm from '../appointment-forms/create-appointment-form.component';
 import { SearchTypes } from '../types';
 
 const PatientSearch: React.FC = () => {
@@ -27,8 +26,6 @@ const PatientSearch: React.FC = () => {
           <SearchResults patients={[]} toggleSearchType={toggleSearchType} />
         ) : searchType === SearchTypes.SCHEDULED_VISITS ? (
           <PatientScheduledVisits toggleSearchType={toggleSearchType} patient={selectedPatient} />
-        ) : searchType === SearchTypes.FORM ? (
-          <PatientForm patient={selectedPatient} patientUuid={selectedPatient.id} />
         ) : null}
       </div>
     </>
